Move file and bug fix
